100% done

+++ src/CustomerPackage/Feedback.java
@@ -4,16 +4,17 @@ public class Feedback {
     private final String feedbackId;
     private final String orderId;
     private final String customerId;
-    private final int rating;
+    private final double rating;
     private final String comment;
 
-    public Feedback(String feedbackId, String orderId, String customerId, int rating, String comment) {
+    public Feedback(String feedbackId, String orderId, String customerId, double rating, String comment) {
         this.feedbackId = feedbackId;
         this.orderId    = orderId;
         this.customerId = customerId;
         this.rating     = rating;
         this.comment    = comment != null ? comment : "";
     }
+
     public String toCSV() {
         String safeComment = comment.replace("\n", " ").replace(",", ";");
         return String.join(",",
@@ -24,43 +25,30 @@ public class Feedback {
                 safeComment
         );
     }
+
     public static Feedback fromCSV(String csv) {
         if (csv == null || csv.isEmpty()) return null;
         String[] parts = csv.split(",", 5);
         if (parts.length < 5) return null;
-        String id    = parts[0].trim();
-        String oid   = parts[1].trim();
-        String cid   = parts[2].trim();
-        int rate;
+        String id  = parts[0].trim();
+        String oid = parts[1].trim();
+        String cid = parts[2].trim();
+        double rate;
         try {
-            rate = Integer.parseInt(parts[3].trim());
+            rate = Double.parseDouble(parts[3].trim());
         } catch (NumberFormatException e) {
             return null;
         }
-        String msg   = parts[4].trim().replace(";", ",");
+        String msg = parts[4].trim().replace(";", ",");
         return new Feedback(id, oid, cid, rate, msg);
     }
 
     // Getters
-    public String getFeedbackId() {
-        return feedbackId;
-    }
-
-    public String getOrderId() {
-        return orderId;
-    }
-
-    public String getCustomerId() {
-        return customerId;
-    }
-
-    public int getRating() {
-        return rating;
-    }
-
-    public String getComment() {
-        return comment;
-    }
+    public String getFeedbackId() { return feedbackId; }
+    public String getOrderId()    { return orderId; }
+    public String getCustomerId() { return customerId; }
+    public double getRating()     { return rating; }
+    public String getComment()    { return comment; }
 
     @Override
     public String toString() {
